+++ schemaorggo/comment.go
@@ -10,15 +10,16 @@ type Comment struct {
 
 	// DownvoteCount see : https://schema.org/downvoteCount
 	// The number of downvotes this question, answer or comment has received from the community.
-	DownvoteCount int `json:"downvoteCount,omitempty"`
+	DownvoteCount float64 `json:"downvoteCount,omitempty"` // types : Integer
 
 	// ParentItem see : https://schema.org/parentItem
 	// The parent of a question, answer or item in general.
-	ParentItem *Question `json:"parentItem,omitempty"`
+	ParentItem *Question `json:"parentItem,omitempty"` // types : Question
 
 	// UpvoteCount see : https://schema.org/upvoteCount
 	// The number of upvotes this question, answer or comment has received from the community.
-	UpvoteCount int `json:"upvoteCount,omitempty"`
+	UpvoteCount float64 `json:"upvoteCount,omitempty"` // types : Integer
+
 }
 
 func (v Comment) MarshalJSONWithTypeContext() ([]byte, error) {
