+++ schemaorggo/itemlist.go
@@ -22,7 +22,8 @@ type ItemList struct {
 
 	// NumberOfItems see : https://schema.org/numberOfItems
 	// The number of items in an ItemList. Note that some descriptions might not fully describe all items in a list (e.g., multi-page pagination); in such cases, the numberOfItems would be for the entire list.
-	NumberOfItems int `json:"numberOfItems,omitempty"`
+	NumberOfItems float64 `json:"numberOfItems,omitempty"` // types : Integer
+
 }
 
 func (v ItemList) MarshalJSONWithTypeContext() ([]byte, error) {
