+++ schemaorggo/question.go
@@ -10,23 +10,24 @@ type Question struct {
 
 	// AcceptedAnswer see : https://schema.org/acceptedAnswer
 	// The answer that has been accepted as best, typically on a Question/Answer site. Sites vary in their selection mechanisms, e.g. drawing on community opinion and/or the view of the Question author.
-	AcceptedAnswer *Answer `json:"acceptedAnswer,omitempty"`
+	AcceptedAnswer *Answer `json:"acceptedAnswer,omitempty"` // types : Answer
 
 	// AnswerCount see : https://schema.org/answerCount
 	// The number of answers this question has received.
-	AnswerCount int `json:"answerCount,omitempty"`
+	AnswerCount float64 `json:"answerCount,omitempty"` // types : Integer
 
 	// DownvoteCount see : https://schema.org/downvoteCount
 	// The number of downvotes this question, answer or comment has received from the community.
-	DownvoteCount int `json:"downvoteCount,omitempty"`
+	DownvoteCount float64 `json:"downvoteCount,omitempty"` // types : Integer
 
 	// SuggestedAnswer see : https://schema.org/suggestedAnswer
 	// An answer (possibly one of several, possibly incorrect) to a Question, e.g. on a Question/Answer site.
-	SuggestedAnswer *Answer `json:"suggestedAnswer,omitempty"`
+	SuggestedAnswer *Answer `json:"suggestedAnswer,omitempty"` // types : Answer
 
 	// UpvoteCount see : https://schema.org/upvoteCount
 	// The number of upvotes this question, answer or comment has received from the community.
-	UpvoteCount int `json:"upvoteCount,omitempty"`
+	UpvoteCount float64 `json:"upvoteCount,omitempty"` // types : Integer
+
 }
 
 func (v Question) MarshalJSONWithTypeContext() ([]byte, error) {
